refactor(frontend): extract parseA2AError into its own util
Move the JSON-RPC error parser and its five regex constants out of
use-message-streaming.ts into chat/utils/parse-a2a-error.ts. The hook
imports from the new module; the existing describe('parseA2AError', ...)
suite in use-message-streaming.test.ts updates its import path only.
Function signature and behavior are unchanged.
